third_party/blink/web_tests/platform/generic/storage/indexeddb/mozilla/objectstorenames-expected.txt

Test IndexedDB's objectStoreNames array

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

dbname = "objectstorenames.html"
indexedDB.deleteDatabase(dbname)
indexedDB.open(dbname)
objectStore = db.createObjectStore(objectStoreName, { keyPath: 'id', autoIncrement: true });
PASS 'objectStoreNames' in db is true
PASS db.objectStoreNames.length is 1
PASS db.objectStoreNames.item(0) is objectStoreName
PASS successfullyParsed is true

TEST COMPLETE
